Patch new sistr database
The sistr developers have provided a new updated database for use with
sistr version 1.0.2 which cleans up some of the issues identified by
us and others. The new database is provided in a branch of the of the
sistr GitHub repo.

Here, I use SVN to download that specific directory of the repo.

Relates to issue #7