Implement proper %test section
The original implementation only tested sistr, which was missing
testing the wrapper CLI.

This is now designed to work separately from building the container.
This is because building singularity containers with a recipe requires
sudo, but we want to make sure tests run under a normal server user. So,
containers should be built with:

sudo singularity buid -T image.simg Sigularity \
    && singularity test image.simg

This approach ensures that the container will work under a normal user
permission environment, and should help us catch any issues similar
to those outlined in #22.