Implement echo of DB_UPDATE env var in %runscript
This relates to commmit 377bc0, and changes how we were creating
environment variables that should be available
when the image is mounted in the %post section of the Singularity recipe.
Singularity docs recommends using appending the file referened by the
SINGULARITY_ENVIRONMENT variable in the %post section. This file is
sourced whenever the image is mounted.

In this case, we use it to identify the date the DB was last updated.