Remove hard coding of contigs.fa filename
In the original implementation, we expected contig files to be named
contigs.fa. This limited our ability to use assemblies with other names
(e.g., spades.fa).

This implementation expects the user to provide a path to the assembly,
which can be called anything. The path is passed on to the sistr
command without modification.