Fix issue #22
As sistr needs a writable directory to extract its database to, this
implementation makes sure every call to sistr happens in a temp
directory. This may come with a small performance penalty, as we are
continuously extracting the DB. But, it gives us stability which is
what we really want. Hopefully, that penalty is minimised by using
fast scratch disk for temp.

Note that here we have to manually patch in the new database in to the
extracted folder.